+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - On `ChessGame.jumpToMainlinePly`, changed `MoveNotFoundException` to
   `IllegalArgumentException` when the target ply is negative
 - On `UCIEngineWrapper.startAnalysisSync`, changed function result to `AnalysisResult` including `EngineLine` data and best move lan string.
+- On `SyzygyTablebase` and `GaviotaTablebase`, added checking the constructor's given directory exists / is readable / has any of syzygy/gaviota file 
 
 ### Fixed
 - Fixed not showing `csl`, `cal` annotation when exporting pgn.